Quick launch of the funtion
