+++ csv-exporter/package.json
@@ -13,6 +13,7 @@
     "dotenv": "^16.0.0",
     "eslint": "^8.10.0",
     "node-fetch": "^3.2.0",
-    "objects-to-csv": "^1.3.6"
+    "objects-to-csv": "^1.3.6",
+    "serverless": "^3.12.0"
   }
 }
